requirements.txt
Django
djangorestframework
drf-spectacular
Pillow
lxml
beautifulsoup4
requests
reportlab
Faker
